rename kungfurun -> kungfu/runner
