+++ srcs/go/cmd/kungfu-modelarts-launcher/kungfu-modelarts-launcher.go
@@ -5,16 +5,16 @@ import (
 	"time"
 
 	"github.com/lsds/KungFu/srcs/go/job"
-	run "github.com/lsds/KungFu/srcs/go/kungfurun"
+	"github.com/lsds/KungFu/srcs/go/kungfu/runner"
 	"github.com/lsds/KungFu/srcs/go/log"
 	"github.com/lsds/KungFu/srcs/go/platforms/modelarts"
 	"github.com/lsds/KungFu/srcs/go/utils"
-	runner "github.com/lsds/KungFu/srcs/go/utils/runner/local"
+	"github.com/lsds/KungFu/srcs/go/utils/runner/local"
 )
 
-var f run.FlagSet
+var f runner.FlagSet
 
-func init() { run.Init(&f) }
+func init() { runner.Init(&f) }
 
 func main() {
 	t0 := time.Now()
@@ -38,7 +38,7 @@ func main() {
 		ctx, cancel = context.WithTimeout(ctx, f.Timeout)
 		defer cancel()
 	}
-	if err := runner.RunAll(ctx, procs, f.VerboseLog); err != nil {
+	if err := local.RunAll(ctx, procs, f.VerboseLog); err != nil {
 		utils.ExitErr(err)
 	}
 }
