+++ srcs/go/kungfu/runner/flags_test.go
@@ -1,4 +1,4 @@
-package kungfurun
+package runner
 
 import (
 	"flag"
