+++ srcs/go/kungfu/runner/handler.go
@@ -1,4 +1,4 @@
-package kungfurun
+package runner
 
 import (
 	"bytes"
