+++ srcs/go/kungfu/runner/peerspec_test.go
@@ -1,4 +1,4 @@
-package kungfurun
+package runner
 
 import "testing"
