+++ srcs/go/kungfu/runner/simple.go
@@ -1,4 +1,4 @@
-package kungfurun
+package runner
 
 import (
 	"context"
@@ -7,13 +7,13 @@ import (
 	"github.com/lsds/KungFu/srcs/go/log"
 	"github.com/lsds/KungFu/srcs/go/plan"
 	"github.com/lsds/KungFu/srcs/go/utils"
-	runner "github.com/lsds/KungFu/srcs/go/utils/runner/local"
+	"github.com/lsds/KungFu/srcs/go/utils/runner/local"
 )
 
 func SimpleRun(ctx context.Context, selfIPv4 uint32, pl plan.PeerList, j job.Job, verboseLog bool) {
 	procs := j.CreateProcs(pl, selfIPv4)
 	log.Infof("will parallel run %d instances of %s with %q", len(procs), j.Prog, j.Args)
-	d, err := utils.Measure(func() error { return runner.RunAll(ctx, procs, verboseLog) })
+	d, err := utils.Measure(func() error { return local.RunAll(ctx, procs, verboseLog) })
 	log.Infof("all %d/%d local peers finished, took %s", len(procs), len(pl), d)
 	if err != nil {
 		utils.ExitErr(err)
